fix(ADA-1745) - Include modal title into close button aria-label  (#84)
This solves https://kaltura.atlassian.net/browse/ADA-1745
It is related to https://github.com/kaltura/playkit-js-ui/pull/984 and
https://github.com/kaltura/playkit-js-info/pull/112

### Description of the Changes

Please add a detailed description of the change, weather it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

### CheckLists

- [ ] changes have been done against master branch, and PR does not
conflict
- [ ] new unit / functional tests have been added (whenever applicable)
- [ ] test are passing in local environment
- [ ] Travis tests are passing (or test results are not worse than on
master branch :))
- [ ] Docs have been updated